ADD many things.. about review  - 리뷰를 썼을 때 specialty를 center의 specialty에도 추가.  - 리뷰를 썼을 때 매긴 별점으로 평균 별점 계산해서 center에 기록.  - review get할 때 specialty를 잘못 가져와서 수정.  - 센터에서 preference를 수정할 때 리뷰에 있는 specialty는 지우지 않도록 수정.  - center의 rateAvg의 type이 제한되어 있어서 평균 별점이 정확하지 않아서 제한을 넓히도록 수정.

+++ db/models/center.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
       roadAddressName: DataTypes.STRING,
       phone: DataTypes.STRING,
       rateAvg: {
-        type: DataTypes.FLOAT(2, 1),
+        type: DataTypes.FLOAT,
         defaultValue: 0,
         allowNull: false,
       },
